+++ NEWS.md
@@ -1,7 +1,9 @@
 # News about the `tcgsaseq` R package
 
+### Main changes in Version 1.7.2 (2017-07-18):
+* faster permutations
 
-### Main changes in Version 1.7.0 (2017-07-07):
+### Main changes in Version 1.7.1 (2017-07-07):
 * faster implementation
 
 ### Main changes in Version 1.6.5 (2017-12-24) --- *this is only a minor release*: